+++ scripts/macros/characterPaths/bhalgrim.js
@@ -1,10 +1,10 @@
 ﻿import {ggHelpers} from '../../helperFunctions.js';
 
-async function richochetAttack({speaker, actor, token, character, item, args}) {
-	const lastArg = args[args.length-1];
-	if (lastArg.itemData.system.actionType !== 'mwak') return;
-	if( lastArg.hitTargets.length > 0 ) {
-		let target = canvas.tokens.get(lastArg.hitTargets[0].id ?? args[0].hitTargets[0]._id);
+async function richochetAttack({speaker, actor, token, character, item, args, scope, workflow}) {
+	console.log(item);
+	if (item.system.actionType !== 'mwak') return;
+	if( workflow.hitTargets.size > 0 ) {
+		let target = canvas.tokens.get(workflow.hitTargets.first().id ?? args[0].hitTargets[0]._id);
 		let nearbyTargets = ggHelpers.findNearby(target, 10, 'nonHostile');
 		if (nearbyTargets.length < 1) return;
 		let buttons = [{label: 'Richochet Attack', value: true},{label: 'No Richochet', value: false}];
@@ -15,19 +15,28 @@ async function richochetAttack({speaker, actor, token, character, item, args}) {
 			if (!splashTargetId) return;
 			const config = {};
 			const options = { targetUuids: [splashTargetId], showFullCard: false, createWorkflow: true, versatile: false, configureDialog: false };
-			const tempItem = item.clone({ "name": 'Richochet' });
+			const tempItem = item.clone({ "name": 'Richochet', "flags.autoanimations.isEnabled": false });
 			tempItem.flags["midi-qol"].onUseMacroParts.items = tempItem.flags["midi-qol"].onUseMacroParts.items.slice(1);
 			await MidiQOL.completeItemUse(tempItem, config, options);
-			let origin = canvas.tokens.get(lastArg.hitTargets[0].id ?? args[0].hitTargets[0]._id);
+			// let origin = canvas.tokens.get(workflow.hitTargets.first().id ?? args[0].hitTargets[0]._id);
 			let splashTarget = await fromUuid(splashTargetId);
 			if (game.modules.get("sequencer")?.active) {
-				const animation = Sequencer.Database.entryExists("jb2a.hammer.throw");
+				let animFile;
+				switch (workflow.defaultDamageType) {
+					case 'slashing':
+						animFile = 'jb2a.handaxe.throw.01';
+						break;
+					case 'bludgeoning':
+					default:
+						animFile = 'jb2a.hammer.throw';
+				}
+				const animation = Sequencer.Database.entryExists(animFile);
 				if (animation) {
 					new Sequence()
 						.effect()
-						.atLocation(origin)
+						.atLocation(target)
 						.stretchTo(splashTarget)
-						.file("jb2a.hammer.throw")
+						.file(animFile)
 						.repeats(1, 200, 300)
 						.randomizeMirrorY()
 						.play();
@@ -37,13 +46,13 @@ async function richochetAttack({speaker, actor, token, character, item, args}) {
 	}
 }
 
-async function richochetDamage({speaker, actor, token, character, item, args}) {
+async function richochetDamage({speaker, actor, token, character, item, args, scope, workflow}) {
 	if (item.name === 'Richochet') {
-		this.damageItem.hpDamage = Math.min(Math.floor(this.damageItem.appliedDamage/2),this.damageItem.hpDamage);
+		workflow.damageItem.hpDamage = Math.min(Math.floor(workflow.damageItem.appliedDamage/2),workflow.damageItem.hpDamage);
 	}
 }
 
-async function mount({speaker, actor, token, character, item, args}) {
+async function mount({speaker, actor, token, character, item, args, scope, workflow}) {
 	let updates = {
 		'token': {
 			'texture': {