+++ scripts/module.js
@@ -1,5 +1,5 @@
 import { registerSettings } from './settings.js';
-import {macros, setupWorldMacros, setupMacroFolder} from './macros.js';
+import {macros} from './macros.js';
 import {effects} from './effects.js';
 import {ggHelpers} from './helperFunctions.js';
 
@@ -9,8 +9,8 @@ Hooks.once('init', async function() {
 });
 Hooks.once('ready', async function() {
 	if (game.user.isGM) {
-		await setupMacroFolder();
-		await setupWorldMacros();
+		// await setupMacroFolder();
+		// await setupWorldMacros();
 		await ggHelpers.createWorldActors();
 	}
 	if (game.settings.get('garhis-grotto', 'Rage Automation')) { Hooks.on('midi-qol.RollComplete', macros.features.class.barbarian.handleRoll);