+++ changelog.md
@@ -1,24 +1,46 @@
 # ToDo List
-## Warlord Stuff
-* Heroic Strike prompt on hit
-* Urgent Orders ask how many dice to spend
+## CPR Alternates
+* Divine Smite
+* Divine Strike
 
 ## Trapper Ranger
+* Feature Descriptions/functionality
 
 ## Misc
 * Electric Arc
-* Kibbles Humanoid Loot Table
+
 # Changelog
 # 0.1.17
+## **BREAKING**
+The Following must be replaced on the character sheet by fresh versions from the compendium(s)
+* Sneak Attack - Rogue Class Feature
+* Twilight Sanctuary - Twilight Cleric Feature
+* Clockwork Bolt - Cantrip
+* Booming Blade - Cantrip
+* Green Flame Blade - Cantrip
+* Auto Reckless - Optional Barbarian Feature
 ## Fixes
 * More v11 updates
 * Fixed Hex/Hunter's Mark
 * Fixed Hunger of Hadar
 * Fixed Twilight Sanctuary
+## Changes
+* Sneak Attack - refactor to no longer rely on a world macro
+* Booming Blade - refactor to act as a Damage Bonus Macro so it has it's own entry on the chat card, removed reliance on world macros, removed different Potent vs. Non-Potent versions
+* Green Flame Blade - similar refactor to Booming Blade
+* Clockwork Bolt - similar refactor to Booming Blade
+* Auto Reckless - refactor to remove world macro
 ## New
 * Added Flaming Sphere
 * Updated Cloud of Daggers and Flaming Sphere to support picking a damage type to support scribes wizards, no validation checking to filter the possible choices at this time.
 * Rallying Mark Automation for Warlord
+* Heroic Strike Automation for Warlord
+* Urgent Orders/Helpful Word resource spending
+## Character Paths
+* All non-played character's path abilities should be considered broken at this stage, most won't take much work to fix but not a priority unless someone starts using a similar ability.
+* Fixed Bhalgrim's Richochet Attack
+## Awakend Weapons
+* Should be considered entirely obsolete and broken, kept around for inspirational purposes, Kivi's Symbiote was an especially cool design imo.
 
 # 0.0.14
 * Added Clockwork Bolt